[#47][be][routine]style: 클래스 inner 클래스로 변경
